FUUUUUUUUUUUUUUUU

+++ hello_world/6-size.c
@@ -8,11 +8,10 @@
  */
 int main(void)
 {
-
-	printf("Size of char: %zu byte(s)\n", (unsigned long)sizeof(char));
-	printf("Size of int: %zu byte(s)\n", (unsigned long)sizeof(int));
-	printf("Size of long int: %zu byte(s)\n", (unsigned long)sizeof(long int));
-	printf("Size of long long int: %zu byte(s)\n", (unsigned long)sizeof(long long int));
-	printf("Size of float: %zu byte(s)\n", (unsigned long)sizeof(float));
-	return (0);
+printf("Size of char: %zu byte(s)\n", sizeof(char));
+printf("Size of int: %zu byte(s)\n", sizeof(int));
+printf("Size of long int: %zu byte(s)\n", sizeof(long int));
+printf("Size of long long int: %zu byte(s)\n", sizeof(long long int));
+printf("Size of float: %zu byte(s)\n", sizeof(float));
+return (0);
 }
